requirements.txt
flask==3.1.0
groq==0.13.1
python-dotenv==1.0.1
